+++ jabgui/src/main/java/org/jabref/gui/JabRefDialogService.java
@@ -13,7 +13,6 @@ import java.util.function.Consumer;
 
 import javafx.application.Platform;
 import javafx.concurrent.Task;
-import javafx.geometry.Pos;
 import javafx.print.PrinterJob;
 import javafx.scene.Group;
 import javafx.scene.Node;
@@ -27,6 +26,7 @@ import javafx.scene.control.ComboBox;
 import javafx.scene.control.Dialog;
 import javafx.scene.control.DialogPane;
 import javafx.scene.control.Label;
+import javafx.scene.control.ProgressBar;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextInputDialog;
 import javafx.scene.control.Tooltip;
@@ -37,10 +37,8 @@ import javafx.stage.DirectoryChooser;
 import javafx.stage.FileChooser;
 import javafx.stage.Stage;
 import javafx.stage.Window;
-import javafx.util.Duration;
 import javafx.util.StringConverter;
 
-import org.jabref.gui.help.ErrorConsoleAction;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.theme.ThemeManager;
 import org.jabref.gui.util.BaseDialog;
@@ -53,10 +51,14 @@ import org.jabref.logic.importer.FetcherClientException;
 import org.jabref.logic.importer.FetcherException;
 import org.jabref.logic.importer.FetcherServerException;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.strings.StringUtil;
 import org.jabref.model.http.SimpleHttpResponse;
 
+import com.dlsc.gemsfx.infocenter.Notification;
+import com.dlsc.gemsfx.infocenter.NotificationAction;
+import com.dlsc.gemsfx.infocenter.NotificationGroup;
+import com.dlsc.gemsfx.infocenter.NotificationView;
 import com.tobiasdiez.easybind.EasyBind;
-import org.controlsfx.control.Notifications;
 import org.controlsfx.control.TaskProgressView;
 import org.controlsfx.control.textfield.CustomPasswordField;
 import org.controlsfx.dialog.ExceptionDialog;
@@ -75,8 +77,16 @@ public class JabRefDialogService implements DialogService {
     // Snackbar dialog maximum size
     public static final int DIALOG_SIZE_LIMIT = 300;
 
-    private static final Duration TOAST_MESSAGE_DISPLAY_TIME = Duration.millis(3000);
     private static final Logger LOGGER = LoggerFactory.getLogger(JabRefDialogService.class);
+
+    private final NotificationGroup<Path, FileNotification> fileNotifications = new NotificationGroup<>(Localization.lang("Files"));
+    private final NotificationGroup<Object, PreviewNotification> previewNotifications = new NotificationGroup<>(Localization.lang("Preview"));
+    private final NotificationGroup<Object, Notification<Object>> undefinedNotifications = new NotificationGroup<>(Localization.lang("Notifications"));
+    private final NotificationGroup<Task<?>, TaskNotification> taskNotifications = new NotificationGroup<>(Localization.lang("Tasks")) {
+        {
+            setViewFactory(TaskNotificationView::new);
+        }
+    };
 
     private final Window mainWindow;
 
@@ -442,25 +452,23 @@ public class JabRefDialogService implements DialogService {
         // TODO: Change to a notification overview instead of event log when that is available.
         //       The event log is not that user friendly (different purpose).
         LOGGER.debug(message);
+        UiTaskExecutor.runInJavaFXThread(() -> notify(new UndefinedNotification("Info", message)));
+    }
 
-        UiTaskExecutor.runInJavaFXThread(() ->
-                Notifications.create()
-                             .text(message)
-                             .position(Pos.BOTTOM_CENTER)
-                             .hideAfter(TOAST_MESSAGE_DISPLAY_TIME)
-                             .owner(mainWindow)
-                             .threshold(5,
-                                     Notifications.create()
-                                                  .title(Localization.lang("Last notification"))
-                                                  .text(
-                                                          "(" + Localization.lang("Check the event log to see all notifications") + ")"
-                                                                  + "\n\n" + message)
-                                                  .onAction(e -> {
-                                                      ErrorConsoleAction ec = new ErrorConsoleAction();
-                                                      ec.execute();
-                                                  }))
-                             .hideCloseButton()
-                             .show());
+    @Override
+    public void notify(Notification<?> notification) {
+        switch (notification) {
+            case FileNotification fileNotification ->
+                    fileNotifications.getNotifications().add(fileNotification);
+            case PreviewNotification previewNotification ->
+                    previewNotifications.getNotifications().add(previewNotification);
+            case TaskNotification taskNotification ->
+                    taskNotifications.getNotifications().add(taskNotification);
+            case UndefinedNotification undefinedNotification ->
+                    undefinedNotifications.getNotifications().add(undefinedNotification);
+            default ->
+                    undefinedNotifications.getNotifications().add(new UndefinedNotification(notification.getTitle(), notification.getSummary()));
+        }
     }
 
     @Override
@@ -551,5 +559,66 @@ public class JabRefDialogService implements DialogService {
             default ->
                     Localization.lang("Something is wrong on JabRef side. Please check the URL and try again.");
         };
+    }
+
+    public static class FileNotification extends Notification<Path> {
+        public FileNotification(String title, String description) {
+            super(title, description);
+            setOnClick(_ -> OnClickBehaviour.HIDE_AND_REMOVE);
+        }
+    }
+
+    public static class UndefinedNotification extends Notification<Object> {
+        public UndefinedNotification(String title, String description) {
+            super(title, description);
+            setOnClick(_ -> OnClickBehaviour.HIDE_AND_REMOVE);
+        }
+    }
+
+    public static class PreviewNotification extends Notification<Object> {
+        public PreviewNotification(String title, String description) {
+            super(title, description);
+            setOnClick(_ -> OnClickBehaviour.HIDE_AND_REMOVE);
+        }
+    }
+
+    public static class TaskNotification extends Notification<Task<?>> {
+        boolean undefinedTask = false;
+
+        public TaskNotification(Task<?> task) {
+            super(task.getTitle(), task.getMessage());
+            setUserObject(task);
+            if (StringUtil.isBlank(task.getTitle())) {
+                setTitle(Localization.lang("Background task"));
+                undefinedTask = true;
+            }
+            setOnClick(_ -> OnClickBehaviour.NONE);
+            task.setOnSucceeded(_ -> {
+                if (undefinedTask) {
+                    UiTaskExecutor.runInJavaFXThread(this::remove);
+                }
+                setOnClick(_ -> OnClickBehaviour.REMOVE);
+            });
+
+            getActions().add(new NotificationAction<>(Localization.lang("Cancel"), _ -> {
+                task.cancel();
+                return OnClickBehaviour.REMOVE;
+            }));
+        }
+    }
+
+    public static class TaskNotificationView extends NotificationView<Task<?>, TaskNotification> {
+        ProgressBar progressBar = new ProgressBar();
+
+        public TaskNotificationView(TaskNotification notification) {
+            super(notification);
+            progressBar.progressProperty().bind(notification.getUserObject().progressProperty());
+            setContent(progressBar);
+            setShowContent(true);
+        }
+    }
+
+    public List<NotificationGroup<?, ? extends Notification<?>>> getNotificationGroups() {
+        return List.of(fileNotifications, undefinedNotifications, previewNotifications, taskNotifications);
     }
 }