+++ jabgui/src/main/java/org/jabref/gui/exporter/ExportCommand.java
@@ -4,19 +4,16 @@ import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Comparator;
 import java.util.List;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import javafx.stage.FileChooser;
-import javafx.util.Duration;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.LibraryTab;
+import org.jabref.gui.JabRefDialogService;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.ActionHelper;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.desktop.os.NativeDesktop;
-import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.preferences.GuiPreferences;
 import org.jabref.gui.util.FileDialogConfiguration;
 import org.jabref.gui.util.FileFilterConverter;
@@ -29,9 +26,8 @@ import org.jabref.logic.util.TaskExecutor;
 import org.jabref.logic.util.io.FileUtil;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
-import org.jabref.model.entry.BibEntryTypesManager;
 
-import org.controlsfx.control.action.Action;
+import com.dlsc.gemsfx.infocenter.NotificationAction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -43,28 +39,22 @@ public class ExportCommand extends SimpleCommand {
     private static final Logger LOGGER = LoggerFactory.getLogger(ExportCommand.class);
 
     private final ExportMethod exportMethod;
-    private final Supplier<LibraryTab> tabSupplier;
     private final StateManager stateManager;
     private final GuiPreferences preferences;
     private final DialogService dialogService;
-    private final BibEntryTypesManager entryTypesManager;
     private final JournalAbbreviationRepository abbreviationRepository;
     private final TaskExecutor taskExecutor;
 
     public ExportCommand(ExportMethod exportMethod,
-                         Supplier<LibraryTab> tabSupplier,
                          StateManager stateManager,
                          DialogService dialogService,
                          GuiPreferences preferences,
-                         BibEntryTypesManager entryTypesManager,
                          JournalAbbreviationRepository abbreviationRepository,
                          TaskExecutor taskExecutor) {
         this.exportMethod = exportMethod;
-        this.tabSupplier = tabSupplier;
         this.stateManager = stateManager;
         this.preferences = preferences;
         this.dialogService = dialogService;
-        this.entryTypesManager = entryTypesManager;
         this.abbreviationRepository = abbreviationRepository;
         this.taskExecutor = taskExecutor;
 
@@ -127,23 +117,9 @@ public class ExportCommand extends SimpleCommand {
                             finEntries,
                             fileDirForDatabase,
                             abbreviationRepository);
-                    return null; // can not use BackgroundTask.wrap(Runnable) because Runnable.run() can't throw Exceptions
+                    return null; // cannot use BackgroundTask.wrap(Runnable) because Runnable.run() can't throw Exceptions
                 })
-                .onSuccess(save -> {
-                    LibraryTab.DatabaseNotification notificationPane = tabSupplier.get().getNotificationPane();
-                    notificationPane.notify(
-                            IconTheme.JabRefIcons.FOLDER.getGraphicNode(),
-                            Localization.lang("Export operation finished successfully."),
-                            List.of(new Action(Localization.lang("Reveal in File Explorer"), event -> {
-                                try {
-                                    NativeDesktop.openFolderAndSelectFile(file, preferences.getExternalApplicationsPreferences(), dialogService);
-                                } catch (IOException e) {
-                                    LOGGER.error("Could not open export folder.", e);
-                                }
-                                notificationPane.hide();
-                            })),
-                            Duration.seconds(5));
-                })
+                .onSuccess(_ -> dialogService.notify(new ExportSuccessNotification(file)))
                 .onFailure(this::handleError)
                 .executeWith(taskExecutor);
     }
@@ -153,5 +129,22 @@ public class ExportCommand extends SimpleCommand {
         dialogService.notify(Localization.lang("Could not save file."));
         // Need to warn the user that saving failed!
         dialogService.showErrorDialogAndWait(Localization.lang("Save library"), Localization.lang("Could not save file."), ex);
+    }
+
+    private class ExportSuccessNotification extends JabRefDialogService.FileNotification {
+        public ExportSuccessNotification(Path path) {
+            super(Localization.lang("Export operation finished successfully."), path.toString());
+
+            NotificationAction<Path> action = new NotificationAction<>(Localization.lang("Reveal"), _ -> {
+                try {
+                    NativeDesktop.openFolderAndSelectFile(path, preferences.getExternalApplicationsPreferences(), dialogService);
+                } catch (IOException e) {
+                    LOGGER.error("Could not open export folder.", e);
+                }
+                return OnClickBehaviour.REMOVE;
+            });
+
+            getActions().add(action);
+        }
     }
 }